+++ brambl-sdk/src/test/scala/co/topl/brambl/builders/AggregationOpsSpec.scala
@@ -1,9 +1,32 @@
 package co.topl.brambl.builders
 
 import cats.implicits.catsSyntaxOptionId
-import co.topl.brambl.syntax.{bigIntAsInt128, valueToQuantitySyntaxOps}
+import co.topl.brambl.syntax.{bigIntAsInt128, valueToQuantitySyntaxOps, valueToTypeIdentifierSyntaxOps}
 
 class AggregationOpsSpec extends TransactionBuilderInterpreterSpecBase {
+
+  test("typeIdentifier grouping") {
+    val testMap = mockValues.groupBy(_.value.typeIdentifier)
+    val expectedMap = Map(
+      value.value.typeIdentifier               -> Seq(value, value.copy()),
+      groupValue.value.typeIdentifier          -> Seq(groupValue, groupValue.copy()),
+      groupValueAlt.value.typeIdentifier       -> Seq(groupValueAlt),
+      seriesValue.value.typeIdentifier         -> Seq(seriesValue, seriesValue.copy()),
+      seriesValueAlt.value.typeIdentifier      -> Seq(seriesValueAlt),
+      assetGroupSeries.value.typeIdentifier    -> Seq(assetGroupSeries, assetGroupSeries.copy()),
+      assetGroupSeriesAlt.value.typeIdentifier -> Seq(assetGroupSeriesAlt),
+      assetGroup.value.typeIdentifier          -> Seq(assetGroup, assetGroup.copy()),
+      assetGroupAlt.value.typeIdentifier       -> Seq(assetGroupAlt),
+      assetSeries.value.typeIdentifier         -> Seq(assetSeries, assetSeries.copy()),
+      assetSeriesAlt.value.typeIdentifier      -> Seq(assetSeriesAlt),
+      assetGroupSeriesAccumulator.value.typeIdentifier -> Seq(
+        assetGroupSeriesAccumulator,
+        assetGroupSeriesAccumulator.copy()
+      ),
+      assetGroupSeriesAccumulatorAlt.value.typeIdentifier -> Seq(assetGroupSeriesAccumulatorAlt)
+    )
+    assertEquals(testMap, expectedMap)
+  }
 
   test("DefaultAggregationOps.aggregate > different types") {
     val input = Seq(groupValue, groupValue, seriesValue, seriesValue).map(_.value)
